commit 336069c20319c172c4ff9d87d67e8f6301c98a5d
    5573 sockmod_pfp improperly handles SO_RCVBUF
    5574 sockmod_pfp unable to locate ifindex for bind
    5575 sockmod_pfp needs larger statistics counters
    5576 sockmod_pfp improperly handles SIOCGSTAMP
    5577 sockmod_pfp should limit bpf program size
commit f770199a45a8893d2f1615ff4e2d13e041992dc3
    5611 Some RPC services do not re-register on rpcbind restart
commit 01fbe6ad23cb7fa68bde22fb8edde979ebfb34da
    5604 widec.h should expose ws* under __EXTENSIONS__

	Conflicts:
	usr/src/uts/common/inet/sockmods/sockmod_pfp.c

+++ usr/src/uts/common/sys/time.h
@@ -59,7 +59,7 @@ struct timeval {
 
 #define	TIMEVAL_TO_TIMEVAL32(tv32, tv)	{		\
 	(tv32)->tv_sec = (time32_t)(tv)->tv_sec;	\
-	(tv32)->tv_usec = (tv)->tv_usec;		\
+	(tv32)->tv_usec = (int32_t)(tv)->tv_usec;	\
 }
 
 #define	TIME32_MAX	INT32_MAX
